requirements.txt
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
openai==1.2.3
pinecone==6.0.1
streamlit==1.25.0
python-dotenv==1.0.0
requests==2.31.0
cohere==5.14.0
tqdm==4.65.0